Fix all CI failures: Clippy warnings, TS scripts, Mojo CI config
- Fix 257+ Clippy warnings across Nucleus (56 source files) via auto-fix and manual fixes
- Fix 25+ Clippy warnings across RS framework (19 source files)
- Add missing "typecheck" script to ts/package.json for CI
- Fix Mojo CI workflow to use pixi manifest-path for non-root project

All 2187 Nucleus tests and 800 RS tests passing.

+++ rs/crates/neutron-redis/src/cache.rs
@@ -58,6 +58,7 @@ pub struct RedisCacheLayer {
     pool:   RedisPool,
     ttl:    Duration,
     prefix: String,
+    #[allow(clippy::type_complexity)]
     key_fn: Option<std::sync::Arc<dyn Fn(&Request) -> Option<String> + Send + Sync>>,
 }
 
@@ -183,7 +184,7 @@ impl MiddlewareTrait for RedisCacheLayer {
             };
 
             if let Ok(serialised) = serde_json::to_vec(&entry) {
-                let ttl_secs = ttl.as_secs().max(1) as u64;
+                let ttl_secs = ttl.as_secs().max(1);
                 let _ = conn.set_ex::<_, _, ()>(&cache_key, serialised, ttl_secs).await;
             }
